feat(frontend): stream file downloads from the file browser
The file explorer and file viewer download actions now delegate to a
shared `downloadFile` helper: over HTTP it navigates to the streaming
`/api/files/download` endpoint (no in-memory buffering), and under WASM
it keeps the base64 pyodide-bridge path.
